Refactor(Issue #3): Moves the token generation functionality from the <List /> component to the <Home /> component.
- Add a new NavLink path to the <Home /> component in the < Footer /> component
- Modify the <Home /> component to include a button that on click generates a unique token if no token exists in local storage
- Modify the <List  /> component to remove token generation functionality
